Update Micronaut to 4.2.0 and fix test
There has been a change in Micronaut to throw an exception if you block the netty threads.

The test in CookieTenantResolverSpec was doing this, so this change adds an ExecuteOn annotation to GatewayController.

It also stops that test being Stepwise (as it makes debugging hard, and is unnecessary)

Also update libs to latests.