Tenant Argument Binder
This adds an api `Tenant`, which represents the abstract notion of a tenant.

Moreover, it adds a RequestFilter to resolve the tenant and set it as a request attribute.

This pull request adds an argument binder that uses the request attribute resolved in the filter. Because of this, a user can bind the tenant as a controller method parameter